Drop OpenVPN support & update service card component
