+++ src/Pages/Authenticated/Dashboard/Services/ServiceHeader.tsx
@@ -4,51 +4,27 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import SecurityIcon from '@material-ui/icons/Security'
-import PublicIcon from '@material-ui/icons/Public'
 import { Chip } from '@material-ui/core'
-
-import { ReactComponent as OpenVpnIcon } from '../../../../assets/images/ovpn-icon.svg'
+import PublicIcon from '@material-ui/icons/Public'
+import SecurityIcon from '@material-ui/icons/Security'
 import { ReactComponent as WireGuardIcon } from '../../../../assets/images/wg-icon.svg'
-import { ServiceType } from '../../../../commons'
 
 import './ServiceHeader.scss'
 
-const icons = (type: ServiceType): unknown => {
-  switch (type) {
-    case ServiceType.OPENVPN:
-      return <OpenVpnIcon />
-    case ServiceType.WIREGUARD:
-      return <WireGuardIcon />
-  }
-  return <WireGuardIcon />
-}
-
-const displayName = (type: ServiceType): string => {
-  switch (type) {
-    case ServiceType.OPENVPN:
-      return ServiceType.OPENVPN
-    case ServiceType.WIREGUARD:
-      return 'WG' // copyright stuff
-  }
-  return 'Unknown'
-}
-
 interface Props {
-  type: ServiceType
   running: boolean
   whitelisted: boolean
 }
 
-const ServiceHeader = ({ type, running, whitelisted }: Props): JSX.Element => {
+const ServiceHeader = ({ running, whitelisted }: Props) => {
   return (
     <div className="service-header">
       <div className="service-header__logo">
-        {icons(type)}
+        <WireGuardIcon />
         <div className={running ? 'status-dot on' : 'status-dot off'} />
       </div>
       <div className="service-header__name">
-        <p className="name">{displayName(type)}</p>
+        <p className="name">WG</p>
         <p className="type">VPN and Data scraping</p>
       </div>
       <div className="flex-grow" />
