+++ src/commons/identity.utils.ts
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 import { Identity, IdentityRegistrationStatus } from 'mysterium-vpn-js'
-import { EMPTY_IDENTITY } from '../redux/selectors'
+import { IDENTITY_EMPTY } from '../constants/instances'
 const { Registered, InProgress, Unregistered, RegistrationError } = IdentityRegistrationStatus
 
 export const isRegistered = (identity: Identity): boolean => {
@@ -25,5 +25,5 @@ export const isRegistrationError = (identity: Identity): boolean => {
 }
 
 export const isEmpty = (identity: Identity): boolean => {
-  return identity.id === EMPTY_IDENTITY.id
+  return identity.id === IDENTITY_EMPTY.id
 }
