+++ src/constants/instances.ts
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { Fees } from 'mysterium-vpn-js'
+import { CurrentPricesResponse, Fees, Identity, IdentityRegistrationStatus } from 'mysterium-vpn-js'
 
 export const TOKENS_EMPTY = Object.freeze({
   wei: '0',
@@ -23,3 +23,25 @@ export const FEES_EMPTY: Fees = Object.freeze({
   decreaseStake: 0,
   decreaseStakeTokens: TOKENS_EMPTY,
 })
+
+export const CURRENT_PRICES_EMPTY: CurrentPricesResponse = Object.freeze({
+  pricePerHour: BigInt(0),
+  pricePerGib: BigInt(0),
+  pricePerHourTokens: TOKENS_EMPTY,
+  pricePerGibTokens: TOKENS_EMPTY,
+})
+
+export const IDENTITY_EMPTY: Identity = {
+  id: '0x',
+  registrationStatus: IdentityRegistrationStatus.RegistrationError,
+  channelAddress: '0x',
+  balance: 0,
+  balanceTokens: TOKENS_EMPTY,
+  earnings: 0,
+  earningsTokens: TOKENS_EMPTY,
+  earningsTotal: 0,
+  earningsTotalTokens: TOKENS_EMPTY,
+  stake: 0,
+  hermesId: '0x',
+  earningsPerHermes: {},
+}
